Rewrite tests that use implementation classes
Except for RecordJava*Processor, which rely on an internal dump
function.
