+++ compiler-plugin/src/test/kotlin/com/google/devtools/ksp/processor/ReferenceElementProcessor.kt
@@ -18,10 +18,9 @@
 
 package com.google.devtools.ksp.processor
 
+import com.google.devtools.ksp.getPropertyDeclarationByName
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.symbol.*
-import com.google.devtools.ksp.symbol.impl.binary.KSTypeReferenceDescriptorImpl
-import com.google.devtools.ksp.symbol.impl.kotlin.KSTypeImpl
 import com.google.devtools.ksp.visitor.KSTopDownVisitor
 
 open class ReferenceElementProcessor: AbstractTestProcessor() {
@@ -36,16 +35,22 @@ open class ReferenceElementProcessor: AbstractTestProcessor() {
             it.accept(collector, references)
         }
 
-        val sortedReferences = references.filter { it.element is KSClassifierReference && it.origin == Origin.KOTLIN }.sortedBy { (it.element as KSClassifierReference).referencedName() }
+        resolver.getPropertyDeclarationByName("z", true)!!.accept(collector, references)
+        resolver.getPropertyDeclarationByName("w", true)!!.accept(collector, references)
+
+        fun refName(it: KSTypeReference) = (it.element as KSClassifierReference).referencedName()
+
+        val sortedReferences = references.filter { it.element is KSClassifierReference && it.origin == Origin.KOTLIN }.sortedBy(::refName)
         for (i in sortedReferences)
             results.add("KSClassifierReferenceImpl: Qualifier of ${i.element} is ${(i.element as KSClassifierReference).qualifier}")
 
-        val descriptorReferences = sortedReferences.map { KSTypeReferenceDescriptorImpl.getCached((it.resolve() as KSTypeImpl).kotlinType, Origin.SYNTHETIC) }
+        // FIXME: References in getters and type arguments are not compared to equal.
+        val descriptorReferences = references.filter { it.element is KSClassifierReference && it.origin == Origin.KOTLIN_LIB }.distinctBy(::refName).sortedBy(::refName)
         for (i in descriptorReferences) {
             results.add("KSClassifierReferenceDescriptorImpl: Qualifier of ${i.element} is ${(i.element as KSClassifierReference).qualifier}")
         }
 
-        val javaReferences = references.filter { it.element is KSClassifierReference && it.origin == Origin.JAVA }.sortedBy { (it.element as KSClassifierReference).referencedName() }
+        val javaReferences = references.filter { it.element is KSClassifierReference && it.origin == Origin.JAVA }.sortedBy(::refName)
         for (i in javaReferences) {
             results.add("KSClassifierReferenceJavaImpl: Qualifier of ${i.element} is ${(i.element as KSClassifierReference).qualifier}")
         }
